Clearer abstraction and hierarchy.
