+++ conf/datacafe.properties
@@ -1,36 +1,7 @@
 #
 #
-# Copyright (c) 2015, Pradeeban Kathiravelu <pradeeban.kathiravelu@tecnico.ulisboa.pt>
+# Copyright (c) 2016, Pradeeban Kathiravelu <pradeeban.kathiravelu@tecnico.ulisboa.pt>
 #
 
-hazelcastXml=conf/hazelcast.xml
-
-# Name of the simulation cluster. Default is "main". Change this for parallel simulations.
-# Uncomment and use different cluster names for multi-tenanted deployments.
-mainCluster="main"
-# mainCluster2="main2"
-# mainCluster3="main3"
-
-# Name of the supportive cluster. Default is "sub". Change this for parallel Supporter clusters.
-# The cluster used for scaling and health decisions.
-# Example use case: IntelligentAdaptiveScaler.
-subCluster="sub"
-
-
-## Health-Monitor properties
-# Mention timeBetweenHealthChecks property along with the params,
-# to enable health checking.
-timeBetweenHealthChecks=10
-## Params
-#highThresholdProcessCpuLoad=0.15
-#lowThresholdProcessCpuLoad=0.02
-#
-#
-#### Auto Scaler properties
-#maxNumberOfInstancesToBeSpawned=3
-###in seconds
-#timeBetweenScalingDecisions=60
-##mode: auto - for scaling up.
-## adaptive - for adding existing instances adaptively into
-## the simulation cluster, based on the load.
-#scalingMode=adaptive
+# This is a sample Data Cafe properties file.
+# Edit the values and/or add more properties accordingly based on your Data Cafe deployment configurations.
