+++ conf/datacafe_hz.properties
@@ -1,0 +1,39 @@
+#
+#
+# Copyright (c) 2015, Pradeeban Kathiravelu <pradeeban.kathiravelu@tecnico.ulisboa.pt>
+#
+
+# This file is given as a sample for creating a Data Cafe Hazelcast cluster.
+# To be able to read these properties, make sure to move these properties to the datacafe.properties file.
+
+hazelcastXml=conf/hazelcast.xml
+
+# Name of the execution cluster. Default is "main". Change this for parallel simulations.
+# Uncomment and use different cluster names for multi-tenanted deployments.
+mainCluster="main"
+# mainCluster2="main2"
+# mainCluster3="main3"
+
+# Name of the supportive cluster. Default is "sub". Change this for parallel Supporter clusters.
+# The cluster used for scaling and health decisions.
+# Example use case: IntelligentAdaptiveScaler.
+subCluster="sub"
+
+
+## Health-Monitor properties
+# Mention timeBetweenHealthChecks property along with the params,
+# to enable health checking.
+timeBetweenHealthChecks=10
+## Params
+#highThresholdProcessCpuLoad=0.15
+#lowThresholdProcessCpuLoad=0.02
+#
+#
+#### Auto Scaler properties
+#maxNumberOfInstancesToBeSpawned=3
+###in seconds
+#timeBetweenScalingDecisions=60
+##mode: auto - for scaling up.
+## adaptive - for adding existing instances adaptively into
+## the execution cluster, based on the load.
+#scalingMode=adaptive
