+++ modules/datacafe-samples/src/main/java/edu/emory/bmi/datacafe/impl/clinical1/data/Slice.java
@@ -15,7 +15,7 @@
  */
 package edu.emory.bmi.datacafe.impl.clinical1.data;
 
-import edu.emory.bmi.datacafe.core.DataSourceBean;
+import edu.emory.bmi.datacafe.interfaces.DataSourceBean;
 import org.jongo.marshall.jackson.oid.MongoId;
 import org.jongo.marshall.jackson.oid.MongoObjectId;
