+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/conf/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Read the Data Cafe config files.
+ */
+package edu.emory.bmi.datacafe.conf;
