+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/core/AbstractExecutorEngine.java
@@ -15,6 +15,7 @@
  */
 package edu.emory.bmi.datacafe.core;
 
+import edu.emory.bmi.datacafe.interfaces.DataSourceBean;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
