+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/hdfs/HdfsConnector.java
@@ -18,7 +18,7 @@ package edu.emory.bmi.datacafe.hdfs;
 import edu.emory.bmi.datacafe.constants.DatacafeConstants;
 import edu.emory.bmi.datacafe.constants.HDFSConstants;
 import edu.emory.bmi.datacafe.core.CoreDataObject;
-import edu.emory.bmi.datacafe.core.WarehouseConnector;
+import edu.emory.bmi.datacafe.interfaces.WarehouseConnector;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
