+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/interfaces/package-info.java
@@ -1,0 +1,4 @@
+/*
+ * Interfaces that needs to be implemented for data sources and beans.
+ */
+package edu.emory.bmi.datacafe.interfaces;
