+++ modules/datacafe-server/src/main/java/edu/emory/bmi/datacafe/mongo/MongoEngine.java
@@ -1,15 +1,23 @@
 /*
- * Copyright (c) 2015 Pradeeban Kathiravelu and others. All rights reserved.
+ * Copyright (c) 2015-2016, Pradeeban Kathiravelu and others. All Rights Reserved.
  *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License v1.0 which accompanies this distribution,
- * and is available at http://www.eclipse.org/legal/epl-v10.html
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 package edu.emory.bmi.datacafe.mongo;
 
 import edu.emory.bmi.datacafe.constants.MongoConstants;
 import edu.emory.bmi.datacafe.core.DataSource;
-import edu.emory.bmi.datacafe.core.DataSourceEngine;
+import edu.emory.bmi.datacafe.core.AbstractDataSourceEngine;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jongo.MongoCollection;
@@ -21,7 +29,7 @@ import java.util.Map;
 /**
  * Core initialization engine of DataCafe with Mongo datasource.
  */
-public class MongoEngine extends DataSourceEngine {
+public class MongoEngine extends AbstractDataSourceEngine {
     private static Logger logger = LogManager.getLogger(MongoEngine.class.getName());
 
     /**
